Fix: userInfo totalWalkingDistance type correction (Int -> Double)

+++ app/src/main/java/com/codehanzoom/greenwalk/model/UserInfoModel.kt
@@ -8,6 +8,6 @@ data class UserInfoResponseBody(
     val totalDonation: Int,
     val totalStep: Int,
     val totalTrashCount: Int,
-    val totalWalkingDistance: Int
+    val totalWalkingDistance: Double
 )
